添加商品图片修改

+++ src/main/java/com/bootdo/xjbg/dao/OrderDao.java
@@ -21,8 +21,8 @@ public interface OrderDao {
 	List<OrderDO> list(Map<String,Object> map);
 	
 	int count(Map<String,Object> map);
-	
-	int save(OrderDO order);
+
+	Long save(OrderDO order);
 	
 	int update(OrderDO order);
